user_auth/django-multi-tenant-nginx/shared/requirements.txt
Django>=4.2,<5
 djangorestframework>=3.15
 PyJWT>=2.9
 python-decouple>=3.8
 mysqlclient>=2.2
 kubernetes>=29.0.0
 celery>=5.4
 redis>=5.0
 django-celery-beat>=2.5.0
 django-redis>=5.2.0  # 添加django-redis依赖
 djangorestframework-simplejwt>=5.3